+++ .prettierignore
@@ -20,3 +20,4 @@ package-lock.json
 logs
 runtime
 infra
+apps/docs/.next
