+++ apps/server/src/modules/integration/integration.controller.ts
@@ -1,25 +1,12 @@
 import { asyncHandler } from '../../shared/asyncHandler'
-import { AppError } from '../../shared/appError'
-import { sendSuccess } from '../../shared/response'
 
-import { createGithubRepoSchema, syncGithubRepoSchema } from './integration.schema'
+import { connectVercelQuerySchema, connectOAuthQuerySchema } from './integration.schema'
 import { integrationsService } from './integration.service'
 
 import type { Request, Response } from 'express'
 
 const connectVercel = asyncHandler(async (req: Request, res: Response) => {
-    const code = req.query.code as string | undefined
-    const teamId = req.query.teamId as string | undefined
-    const configurationId = req.query.configurationId as string | undefined
-    const state = req.query.state as string | undefined
-
-    if (!code) {
-        throw new AppError('no code provided', 400)
-    }
-
-    if (!state) {
-        throw new AppError('no user id provided', 400)
-    }
+    const { code, state, teamId, configurationId } = connectVercelQuerySchema.parse(req.query)
 
     let userId = state
     let redirectPath = '/profile/integrations'
@@ -40,16 +27,7 @@ const connectVercel = asyncHandler(async (req: Request, res: Response) => {
 })
 
 const connectSupabase = asyncHandler(async (req: Request, res: Response) => {
-    const code = req.query.code as string | undefined
-    const userId = req.query.state as string | undefined
-
-    if (!code) {
-        throw new AppError('no code provided', 400)
-    }
-
-    if (!userId) {
-        throw new AppError('no user id provided', 400)
-    }
+    const { code, state: userId } = connectOAuthQuerySchema.parse(req.query)
 
     await integrationsService.connectSupabase({
         userId,
@@ -60,16 +38,7 @@ const connectSupabase = asyncHandler(async (req: Request, res: Response) => {
 })
 
 const connectNotion = asyncHandler(async (req: Request, res: Response) => {
-    const code = req.query.code as string | undefined
-    const userId = req.query.state as string | undefined
-
-    if (!code) {
-        throw new AppError('no code provided', 400)
-    }
-
-    if (!userId) {
-        throw new AppError('no user id provided', 400)
-    }
+    const { code, state: userId } = connectOAuthQuerySchema.parse(req.query)
 
     await integrationsService.connectNotion({
         userId,
@@ -80,16 +49,7 @@ const connectNotion = asyncHandler(async (req: Request, res: Response) => {
 })
 
 const connectGithub = asyncHandler(async (req: Request, res: Response) => {
-    const code = req.query.code as string | undefined
-    const state = req.query.state as string | undefined
-
-    if (!code) {
-        throw new AppError('no code provided', 400)
-    }
-
-    if (!state) {
-        throw new AppError('no user id provided', 400)
-    }
+    const { code, state } = connectOAuthQuerySchema.parse(req.query)
 
     let userId = state
     let redirectPath = '/profile/integrations'
@@ -134,67 +94,9 @@ const connectGithub = asyncHandler(async (req: Request, res: Response) => {
     return res.redirect(`http://localhost:3000${redirectPath}`)
 })
 
-const getUserGithubRepos = asyncHandler(async (req: Request, res: Response) => {
-    const userId = req.user?.userId as string | undefined
-
-    if (!userId) {
-        throw new AppError('unauthorized', 401)
-    }
-
-    const result = await integrationsService.getUserGithubRepos({ userId })
-    return sendSuccess(res, 'repos fetched successfully', result)
-})
-
-const createRepo = asyncHandler(async (req: Request, res: Response) => {
-    const userId = req.user?.userId as string | undefined
-    const projectId = req.params.projectId as string | undefined
-
-    if (!userId) {
-        throw new AppError('unauthorized', 401)
-    }
-
-    if (!projectId) {
-        throw new AppError('project id is required', 400)
-    }
-
-    const parsedBody = createGithubRepoSchema.parse(req.body)
-    const result = await integrationsService.createRepo({
-        userId,
-        projectId,
-        ...parsedBody,
-    })
-
-    return sendSuccess(res, 'repository created and linked successfully', result)
-})
-
-const updateRepo = asyncHandler(async (req: Request, res: Response) => {
-    const userId = req.user?.userId as string | undefined
-    const projectId = req.params.projectId as string | undefined
-
-    if (!userId) {
-        throw new AppError('unauthorized', 401)
-    }
-
-    if (!projectId) {
-        throw new AppError('project id is required', 400)
-    }
-
-    const parsedBody = syncGithubRepoSchema.parse(req.body)
-    const result = await integrationsService.updateRepo({
-        userId,
-        projectId,
-        ...parsedBody,
-    })
-
-    return sendSuccess(res, 'repository synced successfully', result)
-})
-
 export const integrationsController = {
-    getUserGithubRepos,
     connectVercel,
     connectSupabase,
     connectNotion,
     connectGithub,
-    createRepo,
-    updateRepo,
 }
