+++ apps/server/src/modules/integration/integration.types.ts
@@ -1,7 +1,3 @@
-export type ListGithubRepos = {
-    userId: string
-}
-
 export type ConnectVercel = {
     code: string
     userId: string
@@ -23,35 +19,4 @@ export type ConnectGithub = {
     userId: string
     accessToken: string
     username: string
-}
-
-export type GithubRepo = {
-    id: number
-    name: string
-    fullName: string
-    private: boolean
-    defaultBranch: string
-    updatedAt: string
-    htmlUrl: string
-    cloneUrl: string
-    language: string | null
-    description: string | null
-    owner: {
-        login: string
-        avatarUrl: string
-    }
-}
-
-export type CreateRepo = {
-    userId: string
-    projectId: string
-    name: string
-    private: boolean
-    description?: string
-}
-
-export type UpdateRepo = {
-    userId: string
-    projectId: string
-    commitMessage?: string
 }
