+++ apps/server/src/modules/notification/notification.controller.ts
@@ -1,5 +1,5 @@
-import { asyncHandler } from '../../shared/asyncHandler'
 import { AppError } from '../../shared/appError'
+import { asyncHandler } from '../../shared/asyncHandler'
 import { sendSuccess } from '../../shared/response'
 
 import { NotificationParamsSchema } from './notification.schema'
