+++ apps/server/src/modules/platform/platform.repository.ts
@@ -70,4 +70,61 @@ export const platformRepository = {
             },
         })
     },
+
+    async findUserGithubConnection(id: string) {
+        return prisma.user.findUnique({
+            where: { id },
+            select: {
+                githubToken: true,
+                githubUsername: true,
+                githubConnected: true,
+            },
+        })
+    },
+
+    async findProjectByIdAndUser(data: { projectId: string; userId: string }) {
+        const { projectId, userId } = data
+        return prisma.project.findFirst({
+            where: {
+                id: projectId,
+                userId,
+            },
+        })
+    },
+
+    async updateProjectGithub(data: {
+        projectId: string
+        githubRepoName: string
+        githubRepoOwner: string
+        githubRepoUrl: string
+    }) {
+        const { projectId, githubRepoName, githubRepoOwner, githubRepoUrl } = data
+        return prisma.project.update({
+            where: { id: projectId },
+            data: {
+                githubRepoName,
+                githubRepoOwner,
+                githubRepoUrl,
+            },
+        })
+    },
+
+    async findProjectVersionByIdAndProject(data: { versionId: string; projectId: string }) {
+        const { versionId, projectId } = data
+        return prisma.projectVersion.findFirst({
+            where: {
+                id: versionId,
+                projectId,
+            },
+        })
+    },
+
+    async updateProjectSynced(projectId: string) {
+        return prisma.project.update({
+            where: { id: projectId },
+            data: {
+                githubLastSyncedAt: new Date(),
+            },
+        })
+    },
 }
