+++ apps/server/src/modules/platform/platform.types.ts
@@ -33,3 +33,38 @@ export type StreamBuildLogs = {
     deploymentId: string
     res: Response
 }
+
+export type ListGithubRepos = {
+    userId: string
+}
+
+export type GithubRepo = {
+    id: number
+    name: string
+    fullName: string
+    private: boolean
+    defaultBranch: string
+    updatedAt: string
+    htmlUrl: string
+    cloneUrl: string
+    language: string | null
+    description: string | null
+    owner: {
+        login: string
+        avatarUrl: string
+    }
+}
+
+export type CreateRepo = {
+    userId: string
+    projectId: string
+    name: string
+    private: boolean
+    description?: string
+}
+
+export type UpdateRepo = {
+    userId: string
+    projectId: string
+    commitMessage?: string
+}
