+++ apps/server/test/unit/notification.unit.test.ts
@@ -1,4 +1,5 @@
 import { describe, expect, test } from 'bun:test'
+
 import {
     NotificationSchema,
     NotificationParamsSchema,
