+++ apps/server/test/unit/runtime.unit.test.ts
@@ -1,4 +1,5 @@
 import { describe, expect, test } from 'bun:test'
+
 import { startPreviewSchema, previewIdParamSchema } from '../../src/modules/runtime/runtime.schema'
 
 describe('runtime schemas', () => {
