+++ scripts/stop.sh
@@ -2,7 +2,6 @@
 
 set -e
 
-# Resolve project root from script location
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 PROJECT_ROOT="$(cd "$SCRIPT_DIR/.." && pwd)"
